Events: Remove separate input thread & cleanup
Rename input.py to events.py
Rename Input to EventManager (class)

At this point, zygrader is still in a broken state, but the beginnings
of the new window manager and event queue are working.

Because we now use halfdelay(1) for input, we no longer need
a separate input thread. This reduces a lot of code complexity.
