+++ zygrader/data/model.py
@@ -7,7 +7,7 @@ import subprocess
 import tempfile
 import time
 from collections import Iterable
-from zygrader.ui.input import Input
+from zygrader.ui.events import EventManager
 
 from zygrader.config import preferences
 from zygrader.config.shared import SharedData
@@ -320,8 +320,6 @@ class Submission(Iterable):
 
     def do_resume_code(self, process):
         if process:
-            input = ui.get_input()
-            input.take_input.clear()
             curses.endwin()
             SharedData.RUNNING_CODE = True
             process.send_signal(signal.SIGCONT)
@@ -332,7 +330,7 @@ class Submission(Iterable):
         return False
 
     def compile_and_run_code(self, use_gdb):
-        input = ui.get_input()
+        events = ui.get_events()
 
         if self.do_resume_code(SharedData.running_process):
             stopped = self.wait_on_child(SharedData.running_process)
@@ -363,8 +361,7 @@ class Submission(Iterable):
 
         curses.initscr()
         curses.flushinp()
-        input.take_input.set()
-        input.clear_event_queue()
+        events.clear_event_queue()
         curses.doupdate()
         return True
 
@@ -419,9 +416,8 @@ class Submission(Iterable):
         return False
 
     def run_code(self, executable, use_gdb):
-        input = ui.get_input()
-        input.clear_event_queue()
-        Input.take_input.clear()
+        events = ui.get_events()
+        events.clear_event_queue()
         curses.endwin()
 
         print(chr(27) + "[2J", end="")  # Clear the terminal
