+++ zygrader/email_manager.py
@@ -44,8 +44,8 @@ def update_student_list(window: ui.Window,
                         student_list: ui.components.FilteredList):
     """Update the list of students when the locks change"""
     student_list.refresh()
-    input = ui.get_input()
-    input.push_refresh_event()
+    events = ui.get_events()
+    events.push_refresh_event()
 
 
 def watch_students(window: ui.Window, student_list: ui.components.FilteredList):
