+++ zygrader/ui/__init__.py
@@ -1,5 +1,5 @@
 from .window import Window, WinContext
-from .input import Event, GO_BACK, Input
+from .events import Event, EventManager, GO_BACK
 from . import components
 from . import window
 from . import layers
@@ -9,5 +9,6 @@ def get_window() -> Window:
     """Utility function to make getting the Window simpler in code"""
     return Window.get_window()
 
-def get_input() -> Input:
-    return get_window().input
+
+def get_events() -> EventManager:
+    return get_window().events
